requirements.txt
streamlit
pandas
google-api-python-client
google-auth-oauthlib
beautifulsoup4
joblib
scikit-learn
python-whois
numpy